[Feat] Modify App, Report CRUD

See merge request s04-final/s04p31a201!82

+++ backend/TmiProject/src/main/java/com/tmi/repository/AppRepository.java
@@ -5,7 +5,6 @@ import java.util.List;
 import com.tmi.entity.App;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AppRepository extends JpaRepository<App, Integer> {
-
+public interface AppRepository extends JpaRepository<App, Long> {
     List<App> findAllByProjectIdEquals(Long pid);
 }
